IGNITE-22732 Test updates
