+++ modules/calcite/src/test/java/org/apache/ignite/internal/processors/query/calcite/CalciteQueryProcessorTest.java
@@ -47,7 +47,7 @@ import org.apache.ignite.internal.processors.query.QueryEngine;
 import org.apache.ignite.internal.processors.query.calcite.exec.MailboxRegistryImpl;
 import org.apache.ignite.internal.processors.query.calcite.exec.rel.Inbox;
 import org.apache.ignite.internal.processors.query.calcite.exec.rel.Outbox;
-import org.apache.ignite.internal.processors.query.calcite.rules.AbstractInTxTest;
+import org.apache.ignite.internal.processors.query.calcite.rules.AbstractTransactionalSqlTest;
 import org.apache.ignite.internal.processors.query.calcite.schema.IgniteSchema;
 import org.apache.ignite.internal.processors.query.calcite.schema.IgniteTable;
 import org.apache.ignite.internal.processors.query.calcite.util.Commons;
@@ -74,7 +74,7 @@ import static org.junit.Assume.assumeTrue;
  *
  */
 @WithSystemProperty(key = "calcite.debug", value = "false")
-public class CalciteQueryProcessorTest extends AbstractInTxTest {
+public class CalciteQueryProcessorTest extends AbstractTransactionalSqlTest {
     /** */
     private static IgniteEx client;
