+++ modules/calcite/src/test/java/org/apache/ignite/internal/processors/query/calcite/rules/ProjectScanMergeRuleTest.java
@@ -46,7 +46,7 @@ import static org.apache.ignite.internal.processors.query.calcite.rules.OrToUnio
  * sql execution: SELECT t1.f11, t2.f21 FROM T1 t1 INNER JOIN T2 t2 on t1.f11 = t2.f22"
  * need to eleminate all unused coluns and take into account only: f11, f21 and f22 cols.
  */
-public class ProjectScanMergeRuleTest extends AbstractInTxTest {
+public class ProjectScanMergeRuleTest extends AbstractTransactionalSqlTest {
     /** */
     public static final String IDX_CAT_ID = "IDX_CAT_ID";
